pyproject.toml
[project]
name = "nonebot-plugin-update"
description = "用于检测 Nonebot 插件更新的 Nonebot 插件"
authors = [
    {name = "djkcyl",email = "cyl@cyllive.cn"},
]
dependencies = [
    "nonebot2[fastapi,httpx]>=2.0.0",
    "nonebot-adapter-onebot>=2.2.3",
    "nonebot-plugin-localstore>=0.5.0",
    "packaging>=23.1",
    "nonebot-plugin-apscheduler>=0.3.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
dynamic = ["version"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_update/version.py"
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
